requirements.txt
inky[rpi,example-depends]
Pillow
pyyaml
requests
